% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseBandWidth.R
\name{chooseBandWidth}
\alias{chooseBandWidth}
\alias{shinyBW}
\title{Aids in the choice of appropriate bandwidths for kernel density estimation}
\usage{
chooseBandWidth(data, nH = 100, binned = TRUE, bgridsize = c(151, 151),
  store.fhats = TRUE, MAX.fine = 35, MAX.coar = 15, tol.spur = 0,
  plotit = TRUE, ...)

shinyBW()
}
\arguments{
\item{data}{matrix with tSNE coordinates}

\item{nH}{number of different bandwidths for which to precompute kde's, Default: 100}

\item{binned}{logical flag for binned kernel estimation, passed to \code{\link{Hpi}} and \code{\link{kde}}, Default: TRUE}

\item{bgridsize}{vector of binning grid sizes, passed to \code{\link{Hpi}} and \code{\link{kde}}, Default: c(151, 151)}

\item{store.fhats}{logical whether to return list of density estimates, Default: TRUE}

\item{MAX.fine}{maximum number of peaks allowed for "fine" bandwidth suggestion, Default: 35}

\item{MAX.coar}{maximum number of peaks allowed for "coarse" bandwidth suggestion, Default: 15}

\item{tol.spur}{number of 'spurious' peaks to be tolerated in "coarse" bandwidth suggestion (see details), Default: 0}

\item{plotit}{logical whether to plot an overview showing numbers of counted peaks per bandwidth, 
and peak coordinates for three suggested bandwidths, Default: TRUE}

\item{...}{further args passed to \code{\link{Hpi}}}
}
\value{
A list containing elements: 
\describe{
  \item{\code{peaks}}{list of peak coordinates}
  \item{\code{fhats}}{list of \code{\link{kde}} density estimates}
  \item{\code{H}}{plug-in bandwidth matrix as returned by \code{\link{Hpi}}}
  \item{\code{Hfactor}}{vector of correction factors}
  \item{\code{Hindex}}{vector with indices for \code{H1} (fine), \code{H2} (coarse), and \code{H} (plug-in) bandwidths}
}
}
\description{
Computes kernel density estimates (kde's) for a range of bandwidths, 
detects regions of continuously appearing peaks and suggests bandwidths accordingly
}
\details{
Although in most statistical inference the bandwidth parameter is determined by the data, 
  kernel density estimation using plug-in bandwidth selectors does not guarantee to match 
  human-perceived multimodal distributions. For example in very fine-grained tSNE maps,   
  small but potentially relevant subsets might be "smoothed away" in the presence of few 
  dominant modes. In this context, the bandwidth remains a tunable parameter and the choice
  of an appropriate density estimate depends eventually on a "visual criterion". This function 
  tries to aid in the decision by performing following steps: 
\enumerate{
  \item estimate bivariate plug-in bandwidth matrix \code{H} using \code{\link{Hpi}}
  \item detect (\code{\link{PeakFind}}) and count number of local maxima over range of 
        bandwidth matrices (in multiples of \code{H})
  \item remove noisy peaks (via \code{\link{epsKNNee}} and \code{\link{dbscan}}) to improve 
        significant feature detection in the next step 
  \item obtain significant high negative curvature regions on all detected peaks 
        using \code{\link{featureSignif}}
  \item count, label and visualize peaks falling into these regions: whereas the 
        overall number of peaks (in red) is a monotonically decreasing function 
        of the bandwidth, the numbers of continuously appearing local maxima 
        (in green) are rather stable. 
  \item suggest \code{H1} as the largest bandwidth for which the number of continuous 
        peaks is at maximum (usually accompanied by an overall larger number 
        of peaks) and \code{H2} as the largest bandwidth for which the number (nearly) converges 
        with the overall number of local maxima
  \item optionally inspect computed density estimates via \code{\link{shinyBW}} and 
        make a judicious choice.
}
}
\examples{
\dontrun{
set.seed(123)
x <- rbind(matrix(rnorm(500, sd = 0.3), ncol = 2, 
                  dimnames=list(NULL, c("A", "B"))),
           matrix(rnorm(36, mean = 1, sd = 0.3), ncol = 2),
           matrix(rnorm(100, mean = -1, sd = 0.3), ncol = 2))
BW <- chooseBandWidth(x, plotit = TRUE)
BW[3:5]

# interactive selection of a bandwidth, requires global var 'BW' with slot \\code{fhats} present
idx <- shinyBW()
fhat <- BW$fhats[[idx]]
}
}
\references{
K. Shekhar, P. Brodin, M.M. Davis, A.K. Chakraborty (2014), Automatic classification of 
  cellular expression by nonlinear stochastic embedding (ACCENSE), 
  Proceedings of the National Academy of Sciences, \url{http://www.cellaccense.com/oldver.html},
  
  T. Duong, A. Cowling, I. Koch, M.P. Wand (2008), Feature significance for multivariate 
  kernel density estimation. Computational Statistics and Data Analysis, 52, 4225-4242.
}
\seealso{
\code{\link[ks]{Hpi}}
 \code{\link[ks]{kde}}
 \code{\link[dbscan]{dbscan}}
 \code{\link{epsKNNee}}
 \code{\link[feature]{featureSignif}}
 \code{\link[grDevices]{contourLines}}
}
