% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorDens.R
\name{colorDens}
\alias{colorDens}
\title{A color ramp}
\usage{
colorDens(w)
}
\arguments{
\item{w}{numeric vector}
}
\value{
a vector of hex colors
}
\description{
Generates a sequence of MATLAB's jet-like colors
}
\details{
This function maps the input values to [1, 100] and a applies the colorRamp: c("#007FFF", "cyan", "#7FFF7F", "yellow", "#FF7F00", 
    "red", "#7F0000").
}
\examples{
\dontrun{
x <- c(23:234)
image(as.matrix(x), col=colorDens(x))
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}
}
