% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combCompare.R
\name{combCompare}
\alias{combCompare}
\title{\emph{comb}inatorial \emph{Comp}arisons}
\usage{
combCompare(res, sel, data, delta = NA, beta = 0.25, threshold = 0.5)
}
\arguments{
\item{res}{data.frame with column \code{bin.idx}, returned by \code{\link{contourLoc}}}

\item{sel}{row indices to select in res, i.e. the subsets}

\item{data}{matrix containing marker expression values}

\item{delta}{numeric threshold for robust effect size \eqn{\Delta}, Default: NA}

\item{beta}{numeric threshold for non-overlapping quantiles, Default: 0.25}

\item{threshold}{numeric for distance between two distributions, applied at the same scale 
of marker expression, this also avoids unintentional large \eqn{\Delta} values when expression levels are near zero, 
passed to \code{\link{robDelta}} and \code{\link{phenoQuant}}, Default: 0.5}
}
\value{
A list containing elements:
\describe{
  \item{idx}{list of contour/subset indices for pairwise comparisons}
  \item{keep}{logical matrix indicating the results with markers in columns for which subsets have been compared}
  \item{delta}{numeric matrix with computed signed \eqn{\Delta} values, same dimensions as \code{keep}}
  \item{beta}{numeric matrix with signs indicating non-overlap and direction, same dimensions as \code{keep}}
}
}
\description{
compares marker expression levels between each pair of selected subsets
}
\details{
This helper function compares each pair of selected subsets (or contours) for phenotypic similarity for  
 each marker found in \code{data} by means of two dicision boundaries: quantile overlap and/or a minimally   
 required robust effect size \eqn{\Delta}. Since the former does not provide information on how much two subsets 
 differ with regard to a given marker, the latter may be used to estimate the importance of each marker to 
 differentiate between two subsets (as in \code{\link{phenoReduce}}). As possible results, markers that do not
 provide distinction between any two subsets are being excluded from downstream analysis or two phenotypically 
 similar subsets might be merged (as in \code{\link{contourMerge}}).
}
\seealso{
\code{\link{contourLoc}}
 \code{\link{contourMerge}}
 \code{\link{robDelta}}
 \code{\link{exprs2con}}
 \code{\link{findCutoffs}}
 \code{\link{phenoQuant}}
}
\keyword{internal}
