% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourMerge.R
\name{contourMerge}
\alias{contourMerge}
\title{Merge Similar Subsets}
\usage{
contourMerge(con, selected, keep.inner = FALSE, keep.outer = FALSE, minBins,
  data, delta = NA, beta = 0.25, threshold = 0.5)
}
\arguments{
\item{con}{object returned by \code{\link{shrinkSpur}}}

\item{selected}{indices of contours to be compared, useful to manually deselect contours}

\item{keep.inner}{if TRUE, inner contours are not merged even though there are similar (to the outer), Default: FALSE}

\item{keep.outer}{if TRUE, outer contours are not removed even though there are different from inner, Default: FALSE}

\item{minBins}{minimum difference in numbers of observations wihin an outer and inner contours to consider the exclusion (setdifference)
of inner and outer for comparisons (see \code{\link{shrinkSpur}}).}

\item{data}{matrix containing marker expression values}

\item{delta}{numeric threshold for robust effect size \eqn{\Delta}, Default: NA}

\item{beta}{numeric threshold for non-overlapping quantiles, Default: 0.25}

\item{threshold}{numeric for distance (at asinh scale) between two distributions to avoid unintentional 
large \eqn{\Delta} values when expression levels are near zero, Default: 0.5}
}
\value{
\code{con} list object with added fields:
\describe{
  \item{\code{merged}}{indices of kept contours}
  \item{\code{data}}{expression matrix}
  \item{\code{comb}}{\describe{
                       \item{\code{sel}}{data.frame containing indices for each set of inner and outer contours}
                       \item{\code{res}}{list of results for each set of inner and outer comparisons as returned by \code{\link{combCompare}}}
                     }}
}
}
\description{
evaluates phenotypic similarity of contour-gated subsets. An outer contour is removed if it contains 
       inner subsets that have different phenotype(s). Otherwise, inner subsets are "merged" with the outer.
}
\details{
Contour-gated subsets are compared for phenotypic similarity with regard to marker expression provided in \code{data}. 
       They will be considered dissimilar if, for any marker, their symmetric quantiles (as specified by \code{beta}) overlap and/or  
       the expression exceeds the limit for robust effect size \eqn{\Delta}. Comparisons are performed using \code{\link{combCompare}} for  
       all outmost contours and their exclusion (i.e. the "root") as well as for each set of outer and inner contours. The results of the
       final comparisons between dissimilar subsets are returned with the contour list object and may be inspected via the 
       high-level plotting function \code{\link{plotPheno}}. Note that currently no print, show, or summary functions are implemented.
}
\examples{
\dontrun{
 set.seed(123) 
 x <- rbind(matrix(rnorm(100, sd = 0.3), ncol = 2, 
                   dimnames=list(NULL, c("A", "B"))),
            matrix(rnorm(36, mean = 1, sd = 0.3), ncol = 2),
            matrix(rnorm(100, mean = -1, sd = 0.3), ncol = 2))
 BW <- chooseBandWidth(x) 
 
 # get density estimate
 fhat <- BW$fhats[[ BW$Hindex[1] ]]
 
 # get contour list
 cs <- contourLoc(fhat=fhat, offset = 0.01, n.levels = 20, peaks2merge = 5, minBins = 6,
                  keep.level.by = "max.bin.num", plotit = TRUE)
 
 # remove contours containing too few observations
 css <- shrinkSpur(cs) 
 
 # compare and keep outer subsets
 csm <- contourMerge(css, keep.outer=T, keep.inner=F, data=x, 
                     delta=2, beta=0.15, threshold=0.2)
 plotContours(csm, col="grey", pch="+", cex=0.5, lwd=2.5, fhat.levels=T, labels="peak.ids")
 
 # available subsets
 rownames(csm$res)
 
 # sets of outer and inner contours
 csm$comb$sel
 
 # compare within outer 5:
 plotContours(csm, comp.sel=c(1, 4, 8), col="grey", pch="+", cex=0.5, lwd=2.5, fhat.levels=T, 
              labels="peak.ids", filwd=.5, density=50)
 plotPheno(csm, comp.sel=c(1, 4, 8)) # inner differ phenotypically
 
 # outer removed
 csm <- contourMerge(css, keep.outer=F, keep.inner=F, data=x, 
                     delta=2, beta=0.15, threshold=0.2)
 plotContours(csm, col="grey", pch="+", cex=0.5, lwd=2.5, fhat.levels=T, labels="peak.ids")
 
 # show effect size as well
 plotPheno(csm, comp.sel=T, ask=F) 
}
}
\seealso{
\code{\link{combCompare}}
 \code{\link{plotPheno}}
}
