% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffExtract.R
\name{cutoffExtract}
\alias{cutoffExtract}
\title{Extract Gate Boundaries from GatingSet}
\usage{
cutoffExtract(x, yChnl, xyChnl)
}
\arguments{
\item{x}{matrix with gate vertices}

\item{yChnl}{channel against which target marker has been plotted in Cytobank}

\item{xyChnl}{optional channel against which yChnl was plotted in Cytobank}
}
\value{
cutoff value
}
\description{
convenience function
}
