% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuts2Phenotype.R
\name{cuts2Phenotype}
\alias{cuts2Phenotype}
\title{Assign Phenotypes}
\usage{
cuts2Phenotype(con, cutoffs, subsets, beta, verbose = TRUE)
}
\arguments{
\item{con}{object returned by \code{\link{contourMerge}} or \code{\link{findCutoffs}}}

\item{cutoffs}{list of vectors with cutoff-values per marker}

\item{subsets}{indices of subsets to select}

\item{beta}{numeric probability value in [0,1] for symmetric quantiles}

\item{verbose}{flag to print out phenotypes, Default: TRUE}
}
\value{
A list with fields:
\describe{
  \item{\code{phenocode}}{matrix with char encoded phenotypes}
  \item{\code{phenotype}}{phenotypes comprising all markers}
  \item{\code{nonredundant}}{phenotypes excluding markers that show the same phenotype among all subsets}
}
}
\description{
categorizes selected subsets according to 1D 'cutoffs' and assigns phenotype labels
}
\seealso{
\code{\link{phenoCutoff}}
}
