% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsKNNee.R
\name{epsKNNee}
\alias{epsKNNee}
\title{Estimate EPS}
\usage{
epsKNNee(x, k, plotit = FALSE)
}
\arguments{
\item{x}{matrix of observations}

\item{k}{number of nearest neighbors}

\item{plotit}{flag to plot the knee, Default: FALSE}
}
\value{
eps, size of the epsilon neighborhood
}
\description{
robust knee detection to estimate a suitable value for dbscan eps neighborhood
}
\examples{
\dontrun{
# example data from fpc
set.seed(665544)
n <- 100
x <- cbind(
  x = runif(10, 0, 10) + rnorm(n, sd = 0.2),
  y = runif(10, 0, 10) + rnorm(n, sd = 0.2)
  )
epsKNNee(x, k = 3, plotit=TRUE)
}
}
\seealso{
\code{\link[dbscan]{kNNdist}}
 \code{\link[dbscan]{kNNdistplot}}
 \code{\link[robust]{lmRob}}
}
