% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs2con.R
\name{exprs2con}
\alias{exprs2con}
\title{Expression Data to Contours}
\usage{
exprs2con(con, data, markers, clus.res, delta, beta, threshold, dwnsmpl = 512)
}
\arguments{
\item{con}{object returned by \code{\link{contourMerge}} or \code{\link{findCutoffs}}}

\item{data}{expression matrix}

\item{markers}{to be used, vector of names matching columns in \code{data}}

\item{clus.res}{vector containing class assigments, of the same length as number of rows in \code{data}}

\item{delta}{numeric threshold for robust effect size \eqn{\Delta}}

\item{beta}{numeric threshold for non-overlapping quantiles}

\item{threshold}{numeric for distance (at asinh scale) between two distributions to avoid unintentional 
large \eqn{\Delta} values when expression levels are near zero}

\item{dwnsmpl}{number of target events to be sampled from each cluster or contour gate, Default: 512}
}
\value{
A list object with following updated fields:
\describe{
  \item{\code{res}}{\describe{
                       \item{\code{peaks}}{list of indices or class for each contour or cluster}
                       \item{\code{bin.idx}}{list of indices of observations within each contour or cluster}
                     }}
  \item{\code{update}}{indices of contours or clusters, corresponding to rows in \code{res}}
  \item{\code{merged}}{indices of kept contours}
  \item{\code{data}}{downsampled expression matrix}
  \item{\code{comb}}{\describe{
                       \item{\code{sel}}{data.frame containing indices for each set of inner and outer contours}
                       \item{\code{res}}{list of results for subset or cluster comparisons as returned by \code{\link{combCompare}}}
                     }}
  \item{\code{eprs.data}}{}
}
}
\description{
contour-gating of raw expression data or phenotyping of clustering results
}
\details{
This function may be used to update a contour list object to use event-level (instead of bin-level) data or convert results 
      obtained from clustering into a 'contour' object in order to perform phenotypic comparisons of clusters.
}
