% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{innerPolyExclude}
\alias{innerPolyExclude}
\alias{arpoly}
\alias{inpoly}
\alias{inpolyl}
\alias{closest}
\title{Polygonal geometry}
\usage{
innerPolyExclude(outer, inner)

arpoly(poly)

inpoly(points, poly)

inpolyl(points, poly)

closest(query, point, distance = TRUE)
}
\arguments{
\item{outer}{polygon which is a list with elements x and y containing x- and y-coordinates}

\item{inner}{list of inner polygon-lists with elements x and y containing x- and y-coordinates}

\item{poly}{list with elements x and y containing x- and y-coordinates}

\item{points}{numeric matrix with points in rows and coordinates in first two columns}

\item{query}{set of points, a matrix with 2 columns}

\item{point}{vector of length 2}

\item{distance}{logical whether to return the shortest distance. If FALSE, the nearest neighbour index is returned, Default: TRUE}
}
\description{
Internal functions to work with polygons
}
\details{
\code{innerPolyExclude} combines polygons so that inner polygons form an exclusion inside an outer polygon.
         \code{closest} finds the nearest neighbour of a given point from a set of points in 2D.
         The functions \code{arpoly}, \code{inpoly}, and \code{inpolyl} are wrappers around splancs' \code{\link{areapl}},
         \code{\link{inpip}}, and \code{\link{inout}}, respectively.
         \code{inner} usually is a list of contours, both \code{outer} and \code{poly} a contour, as returned by \code{\link{contourLines}}.
}
\seealso{
\code{\link[splancs]{areapl}}
 \code{\link[splancs]{inout}}
 \code{\link[splancs]{inpip}}
}
\keyword{internal}
