% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get1dGates.R
\name{get1dGates}
\alias{get1dGates}
\title{1D Gates}
\usage{
get1dGates(chnl, mrkr, cutoffs, boundaries)
}
\arguments{
\item{chnl}{channel name}

\item{mrkr}{marker name}

\item{cutoffs}{numeric vector of cutoff-values}

\item{boundaries}{a lower and an upper limit of the expression data, e.g. the range}
}
\value{
A list containing the flowCore filter objects
}
\description{
creates 1d flowCore filters for 1, 2, or 3 cutoffs and a given marker
}
\details{
The function creates a set of all possible 1d-\code{\link{rectangleGate}s} for a given number of cutoffs and assigns filterId's according 
         to the phenotyping syntax specified in \code{\link{phenoCutoff}}. Filters of different markers may be combined to create multidimensional 
         'boolean' gates to target specific phenotypes.
}
\examples{
\dontrun{
 #EXAMPLE1
 # in flowCore
 cd3 <- get1dGates("Eu151Di", "CD3", 4, c(0, 8))
 cd4 <- get1dGates("Yb173Di", "CD4", c(2, 5), c(0, 8))
 cd3$`CD3-` * cd4$`CD4+-`
 # add to GatingSet
 exp <- "/CD3- & /CD4+-"
 bf <- boolFilter(exp, filterId = "Mono")
 # add(gs, cd3$`CD3-`)
 # add(gs, cd4$`CD4+-`)
 # add(gs, bf)
}
}
\seealso{
\code{\link[flowCore]{rectangleGate-class}}
 \code{\link{boolFilter}}
}
