% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hingeQuantileCuts.R
\name{hingeQuantileCuts}
\alias{hingeQuantileCuts}
\title{Quantile Gating}
\usage{
hingeQuantileCuts(hings, keep, threshold, find.all = FALSE)
}
\arguments{
\item{hings}{array of dim c(k, 2, 10) with precomputed hinges for a given marker, 
i.e. a sequence of 10 upper and lower quantiles for each of k subsets}

\item{keep}{logical matrix indicating which combination of two subsets 
differ in expression of what marker based on separation thresholds for \code{beta} 
and \code{delta} as a result of \code{\link{combCompare}}}

\item{threshold}{numeric for minimally required distance between two cutoffs}

\item{find.all}{logical flag whether to ignore \code{keep}, Default: FALSE}
}
\value{
numeric vector of cutoff-values
}
\description{
Internal function that calculates cutoffs between lower and upper hinge of any two subsets if possible
}
\details{
This function is used in \code{\link{findCutoffs}}.
}
