% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdeDiff.R
\name{kdeDiff}
\alias{kdeDiff}
\title{Positive Density Difference}
\usage{
kdeDiff(x, fhat, binned = TRUE, bgridsize, H, w0, w1, compute.cont = FALSE)
}
\arguments{
\item{x}{matrix of data values}

\item{fhat}{object returned by \code{\link{kde}}}

\item{binned}{flag for binned estimation, Default: TRUE}

\item{bgridsize}{vector of binning grid sizes}

\item{H}{bandwidth matrix}

\item{w0}{vector of weights for the minuend density, Default is a vector of all ones.}

\item{w1}{vector of weights for the subtrahend density}

\item{compute.cont}{flag for computing 0.1 to 0.99 probability contour levels, Default: FALSE}
}
\value{
object of class \code{\link{kde}} with the positive difference of two density estimates in field \code{fhat}
}
\description{
subtracts two densities and returns the positive difference
}
\seealso{
\code{\link[ks]{kde}}
}
