% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFind.R
\name{PeakFind}
\alias{PeakFind}
\title{Peak detection}
\usage{
PeakFind(fhat, tol = 5e-08)
}
\arguments{
\item{fhat}{object returned by \code{\link{kde}} or \code{\link{kdeDiff}}}

\item{tol}{threshold to avoid detection of noise, Default: 5e-08}
}
\value{
peak points matrix with local densities in 3rd column
}
\description{
Finds peaks in a 2D kernel density estimate
}
\details{
This function identifies local maxima in a kernel density estimate. 
Code has been adaptated from the ACCENSE R-implementation accessible under \url{http://www.cellaccense.com/oldver.html}.
}
\examples{
\dontrun{
library(ks)
x <- rbind(matrix(rnorm(100, sd = 0.3), ncol = 2),
           matrix(rnorm(100, mean = 1, sd = 0.3), ncol = 2))
fhat <- kde(x)
pts <- PeakFind(fhat)
plot(fhat)
points(pts[ , 1:2], pch="*")
}
}
\references{
K. Shekhar, P. Brodin, M.M. Davis, A.K. Chakraborty (2014), Automatic classification of 
  cellular expression by nonlinear stochastic embedding (ACCENSE), 
  Proceedings of the National Academy of Sciences, \url{http://www.cellaccense.com/oldver.html}, 

  E.R. Davies (1997), Machine Vision, volume 609, Academic press New York.
}
\keyword{internal}
