% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoReduce.R
\name{phenoReduce}
\alias{phenoReduce}
\title{Minimal Marker Cutoff Combinations}
\usage{
phenoReduce(con, target.num.marker = NA, b = 0.5)
}
\arguments{
\item{con}{object returned by \code{\link{findCutoffs}}}

\item{target.num.marker}{desired least number of markers, Default: NA}

\item{b}{weighs recall and precision in F-score, Default: 0.5}
}
\value{
\code{con} object with updated and added fields:
\describe{
  \item{\code{cutoffs}}{list with updated cutoffs per marker element}
  \item{\code{targets}}{named indices of subsets that phenotypically differ from at least one other subset}
  \item{\code{reduc}}{\describe{
                 \item{\code{keep}}{logical vector indicating final marker-cutoff-combinations}
                 \item{\code{solutions}}{matrix containing all feasible solutions returned by \code{\link{lp}}}
                 \item{\code{fscore}}{matrix with computed fscores for all feasible solutions}
               }}
}
}
\description{
finds low-dimensional phenotypic representations of target subsets
}
\details{
The function solves an Integer Linear Programm (ILP) to find a set of at most 3 cutoffs and a minimal and/or required 
         number of markers that still allow to discriminate between subsets. In the likely case that multiple solutions are found,
         these are used to gate the target phenotypes and compute F-scores. The optimal solution is selected based on the best average
         accuracy among the subsets. If target.num.marker is specified, the ILP maximizes the sum of effective differences \eqn{Delta}
         between the subsets and a minimal set as well as a solution with at least the required number of markers is returned. See  
         \code{\link{findCutoffs}} for examples. Note that currently no print, show, or summary functions are implemented.
}
\seealso{
\code{\link{findCutoffs}}
 \code{\link[lpSolve]{lp}}
}
