% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoType.R
\name{phenoQuant}
\alias{phenoQuant}
\alias{phenoCutoff}
\title{Quantile Phenotyping}
\usage{
phenoQuant(x, y, beta, th)

phenoCutoff(x, cutoffs, beta)
}
\arguments{
\item{x}{numeric vector of expression values}

\item{y}{numeric vector of expression values}

\item{beta}{numeric probability value in [0,1], threshold for non-overlapping quantiles}

\item{th}{numeric threshold for distance between two distributions, applied at the same scale 
of marker expression}

\item{cutoffs}{numeric vector of cutoff-values}
}
\description{
Internal functions for comparing marker expression levels between two subsets 
             and phenotyping based on symmetric quantiles and cutoffs
}
\details{
\code{phenoQuant} returns a sign indicating non-overlap and direction of two marker distributions.
   \code{phenoCutoff} encodes a subset as a string according to the relative position  
   of its "quantiloid" to given cutoffs (1d gates). Using common syntax, i.e. by 
   combinations of "-", "+", we're able to classify 2 phenotypes per marker (1 cutoff), 
   5 phenotypes per marker (2 cutoffs), or 9 phenotypes per marker (3 cutoffs are rarely used in cytometric immunophenotyping):
\enumerate{
  \item \describe{
          \item{negative}{-}
          \item{positive}{+}
        }
  \item \describe{
          \item{negative}{- -  (below 1st cutoff)}
          \item{low}{-  (below 2nd, spanning 1st)}
          \item{positive}{+  (above 1st, spanning 2nd)}
          \item{dim}{+ -  (between 1st and 2nd)}
          \item{high}{+ +  (above 2nd)}
        }
  \item \describe{
          \item{negative}{- - -  (below 1st)}
          \item{very low}{- -  (below 2nd, spanning 1st)}
          \item{low}{-  (below 3rd, spanning 1st and 2nd)}
          \item{positive}{+  (above 1st, spanning 2nd and 3rd)}
          \item{lowdim}{+ - -  (between 1st and 2nd)}
          \item{dim}{+ -  (between 1st and 3rd)}
          \item{highdim}{+ + -  (between 2nd and 3rd)}
          \item{high}{+ +  (above 2nd, spanning 3rd)}
          \item{very high}{+ + +  (above 3rd)}
        }
 }
}
\keyword{internal}
