% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContours.R
\name{plotContours}
\alias{plotContours}
\title{Contour Plot}
\usage{
plotContours(con, fhat.levels = FALSE, col = "grey", pch = "+",
  cex = 0.5, lwd = 2.5, comp.sel = FALSE, density = NULL, filwd = 1.5,
  labels = c("none", "peak.pts", "peak.ids"), cex.labels = 0.5,
  ask = TRUE, ...)
}
\arguments{
\item{con}{object returned by \code{\link{contourLoc}}, \code{\link{shrinkSpur}}, \code{\link{contourMerge}}, 
\code{\link{findCutoffs}}, or \code{\link{phenoReduce}}}

\item{fhat.levels}{flag to plot contour levels, Default: FALSE}

\item{col}{color for observations points, Default: 'grey'}

\item{pch}{plotting 'character', i.e., symbol to use for observations points, Default: '+'}

\item{cex}{size of observations points, Default: 0.5}

\item{lwd}{width of contour lines, Default: 2.5}

\item{comp.sel}{indices of contours to plot, Default: FALSE}

\item{density}{the density of shading lines, in lines per inch, Default: NULL}

\item{filwd}{width of shading lines, Default: 1.5}

\item{labels}{how peaks shall be plotted, one of "none", "peak.pts", or "peak.ids"}

\item{cex.labels}{size of peak labels, Default: 0.5}

\item{ask}{passed to \code{\link[graphics]{par}}, Default: TRUE}

\item{...}{further args passed to \code{\link[graphics]{plot}}}
}
\description{
high-level plotting function
}
