% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPheno.R
\name{plotPheno}
\alias{plotPheno}
\title{Plot Subset Marker Expression}
\usage{
plotPheno(con, comp.sel = FALSE, marker.ord = NA, relevant.only = FALSE,
  plot.cutoffs = FALSE, delta.panel = TRUE, max.delta.bars = 6,
  ask = TRUE, grobs = FALSE)
}
\arguments{
\item{con}{object returned by \code{\link{contourMerge}}, \code{\link{findCutoffs}}, or \code{\link{phenoReduce}}}

\item{comp.sel}{indices of contour-gated subsets to plot, Default: FALSE}

\item{marker.ord}{order of markers, Default: NA uses alphabetical order}

\item{relevant.only}{flag whether to exclude noninformative markers, Default: FALSE}

\item{plot.cutoffs}{flag to plot horizontal cutoff-lines, Default: FALSE}

\item{delta.panel}{flag whether to plot barcharts for pairwise computed \eqn{Delta} values, Default: TRUE}

\item{max.delta.bars}{number of bars to plot per dodge, Default: 6}

\item{ask}{passed to \code{\link[graphics]{par}}, Default: TRUE}

\item{grobs}{flag whether to return a list of grobs without drawing, see \code{\link[gridExtra]{arrangeGrob}}, Default: FALSE}
}
\description{
high-level plotting function
}
\seealso{
\code{\link{plotContours}}
}
