% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkSpur.R
\name{shrinkSpur}
\alias{shrinkSpur}
\title{Shrink Spurious Contours}
\usage{
shrinkSpur(con, selected, mul = FALSE, minBins, peaks2merge)
}
\arguments{
\item{con}{object returned by \code{\link{contourLoc}}}

\item{selected}{indices of contours corresponding to rows in \code{con$res}, useful to manually select or deselect contours}

\item{mul}{if TRUE, \code{minBins} is multiplied by the number of inner spurious peaks (as there might be multiple) to more 
progressively shrink towards single inner contours, Default: FALSE}

\item{minBins}{minimum tolerated difference in the numbers of observations of an outer and its single inner contour.}

\item{peaks2merge}{may also be lowered in order to remove outmost contours}
}
\value{
\code{con} list object with added field:
\describe{
  \item{\code{shrunk}}{indices of kept contours}
}
}
\description{
removes outer contours that have been reported due to presence of spurious peaks
}
\details{
\code{\link{contourLoc}} might have returned contours that surround a single inner contour (if \code{peaks2merge > 1}), 
       due to the detection of spurious peaks. These contours need to be removed in order to have a sufficient number of observations
       (specified by \code{minBins}) for comparing marker expression levels of contour-gated subsets as performed by \code{\link{contourMerge}}.
       Contours are labeled by tupels of peak indices, so they can be visually identified in a plot (see \code{\link{plotContours}}). Note that 
       currently no print, show, or summary functions are implemented.
}
\examples{
\dontrun{
 set.seed(123) 
 x <- rbind(matrix(rnorm(100, sd = 0.3), ncol = 2, 
                   dimnames=list(NULL, c("A", "B"))),
            matrix(rnorm(36, mean = 1, sd = 0.3), ncol = 2),
            matrix(rnorm(100, mean = -1, sd = 0.3), ncol = 2))
 BW <- chooseBandWidth(x) 
 
 # get density estimate
 fhat <- BW$fhats[[ BW$Hindex[1] ]]
 
 # get contour list
 cs <- contourLoc(fhat=fhat, offset = 0.01, n.levels = 20, peaks2merge = 5, minBins = 6,
                  keep.level.by = "max.bin.num", plotit = FALSE)
 # spurious peak no. 5 
 plotContours(cs, cex.labels=1.5, fhat.levels=T, labels="peak.ids")
 
 # remove outer contours containing too few observations
 css <- shrinkSpur(cs) 
 plotContours(css, fhat.levels=T, labels="peak.ids")
}
}
