% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM.R
\name{sm.test}
\alias{sm.test}
\title{The Skillings-Mack test statistic}
\usage{
sm.test(y, groups = NULL, blocks = NULL, simulate.p.value = FALSE,
  B = 10000, sim.SM = FALSE)
}
\arguments{
\item{y}{matrix with variables in columns and observations in rows}

\item{groups}{a vector or factor object giving the group for the corresponding rows in y, Default: NULL}

\item{blocks}{a vector or factor object giving the blocks or subjects for the corresponding rows in y, Default: NULL}

\item{simulate.p.value}{If TRUE, an estimated p-value based on the Monte Carlo method is calculated. Default: FALSE}

\item{B}{number of permutations, Default: 10000}

\item{sim.SM}{flag whether simulation data should be returned, Default: FALSE}
}
\value{
A list containing elements:
\describe{
  \item{p.value}{vector with p-values}
  \item{statistic}{vector with Skillings-Mack statistics}
  \item{df}{vector with degrees of freedom}
}
If \code{simulate.p.value=TRUE}: a vector of approximate p-values or a list with elements: 
\enumerate{
  \item list with test results based on the chi-squared distribution
  \item vector of Skilings-Mack statistics
  \item matrix containing the permutation distribution of the Skillings-Mack statistic
}
}
\description{
performs a nonparametric Friedmann-type test for incomplete block or repeated measures designs
}
\details{
The Skillings-Mack test statistic is a generalization of the statistic used in Friedman's ANOVA method and in Durbin's rank test.
This nonparametric statistical test is useful for the data obtained from block or repeated measures designs with missing observations  
occurring randomly. The p-values are based on the chi-squared distribution or estimated by Monte Carlo method. 
The latter is recommended for approximating p-values when there are many ties and/or small designs with missing values are conducted.
This function reuses the implementation in \href{https://CRAN.R-project.org/package=Skillings.Mack}{Skillings.Mack} 
authored by Patchanok Srisuradetchai. Code has been slightly modified to allow for simultaneous Monte Carlo permutation of multiple variables,
e.g. to preserve the spatial dependence across a binned histogramm (or any multivariate data structure).
}
\references{
J.H. Skillings and G.A. Mack (1981), On the Use of a Friedman-Type Statistic in Balanced and Unbalanced Block Designs, Technometrics 23(2)
}
\seealso{
\href{https://CRAN.R-project.org/package=Skillings.Mack}{Skillings.Mack}
}
