% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study}
\alias{study}
\title{study meta data.}
\format{A data frame with 36 rows and 11 variables:
\describe{
  \item{Sample}{sample name}
  \item{fcs}{fcs file name}
  \item{Pool}{pooled sample}
  \item{BC}{barcode number}
  \item{Barcode}{barcode sample name}
  \item{Donor}{donor id}
  \item{Region}{brain region, from which samples were isolated}
  \item{Sex}{donor gender}
  \item{Age}{donor age}
  \item{donor.col}{figure color code used for donor}
  \item{region.col}{figure color code used for brain region}
}}
\usage{
study
}
\description{
study meta data.
}
\keyword{datasets}
