% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_names.R
\name{switch.names}
\alias{switch.names}
\title{Switch Channel and Marker Names}
\usage{
switch.names(m, from, to)
}
\arguments{
\item{m}{matrix with alternative names in columns}

\item{from}{vector with names to look up in m}

\item{to}{either one of c("name", "desc", "pretty.desc") or a vector of strings with some target names}
}
\value{
vector of strings
}
\description{
convenience function
}
\details{
columns in \code{m} usually are "name" and "desc" from the \code{\link[flowCore]{parameters}} \code{@Data} slot, i.e.
      the channel names and description (marker names) of a flowFrame. "pretty.desc" could be marker names prettified to not contain any 
      unwanted characters.
}
