% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold2one.R
\name{threshold2one}
\alias{threshold2one}
\title{Scale Weights for Density Difference}
\usage{
threshold2one(w, th)
}
\arguments{
\item{w}{numeric vector, either effect size or -log transformed \emph{p}-values}

\item{th}{numeric threshold, e.g. -log transformed level of significance}
}
\value{
numeric vector of weights summing up to the number of observations
}
\description{
a ramp function
}
\details{
The function piece-wise linear transforms input values such that they are 1 if equal to \code{th} and 
      their sum equals the number of observations.
}
\examples{
\dontrun{
 # vector of p-values
 p <- runif(512)
 w <- threshold2one(-log2(p), -log2(0.05))
 sum(w)
 plot(sort(threshold2one(1:100, 98)))
 plot(sort(threshold2one(1:100, 50)))
 plot(sort(threshold2one(1:100, 2)))
}
}
