% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold2zero.R
\name{threshold2zero}
\alias{threshold2zero}
\title{Scale Weights for Plotting}
\usage{
threshold2zero(w, th)
}
\arguments{
\item{w}{numeric vector}

\item{th}{numeric threshold}
}
\value{
numeric vector
}
\description{
convenience function
}
\details{
Sets values below threshold to zero. Scales weights to sum up to the number of observations.
}
