% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmax}
\alias{tmax}
\title{Tmax statistic.}
\format{A vector of 10,000 Monte Carlo-simulated maximal Skillings-Mack statistics, one for each resampling}
\usage{
tmax
}
\description{
Tmax statistic.
}
\keyword{datasets}
